Modularize EntryCard sub-components like image
