+++ src/EntryCard/EntryCard.jsx
@@ -2,85 +2,14 @@ import {useEffect, useState} from "react";
 import {EntryCardContext} from "../Application/Application.jsx";
 import {useContext} from "react";
 import PropTypes from "prop-types";
-
-function EntryImage(props){
-    const imageClass = "entry-image";
-    const inputId = "change-image";
-    const type = "file";
-    const accept= ".png, .jpg, .jpeg"
-
-    const changeImage = (event) => {
-        if (event.target.files && event.target.files[0]){
-            const file = event.target.files[0]; // first uploaded image
-            const reader = new FileReader();
-            reader.onloadend = () => {props.update(props.index, "image", reader.result);}
-            reader.readAsDataURL(file);
-        }
-    }
-
-    return (
-        <>
-            <img
-                className={imageClass}
-                src={props.image}
-                alt={"Image of " + props.title}
-                onClick={() => document.getElementById(inputId).click()}
-            />
-            <input id={inputId}
-                   type={type}
-                   accept={accept} onChange={changeImage} disabled={props.disabled}/>
-        </>
-    );
-}
-
-function EntryTitle(props){
-    const inputClass = "entry-title"
-    const type = "text";
-
-    return (
-        <>
-            <input
-                className={inputClass}
-                type={type}
-                value={props.title}
-                onChange={(event) => props.update(props.index, "title", event.target.value)}
-                disabled={props.disabled}
-            />
-        </>
-    );
-}
-
-function EntryDate(props){
-    const inputClass = "entry-date"
-    const type = "text";
-
-    return (
-        <>
-            <input
-                className={inputClass}
-                type={type}
-                value={props.date}
-                onChange={(event) => props.update(props.index, "date", event.target.value)}
-                disabled={props.disabled}
-            />
-        </>
-    );
-}
-
-function EntryDescription(props){
-    const inputClass = "entry-description";
-
-    return (
-        <textarea
-            className={inputClass}
-            value={props.description}
-            onChange={(event) => props.update(props.index, "description", event.target.value)}
-            disabled={props.disabled}
-        />
-    );
-}
+import EntryImage from "./EntryCardComponents/EntryImage.jsx";
+import EntryTitle from "./EntryCardComponents/EntryTitle.jsx";
+import EntryDate from "./EntryCardComponents/EntryDate.jsx";
+import EntryDescription from "./EntryCardComponents/EntryDescription.jsx";
 
 function EntryCard(props){
+
+    // Default values for a timeline entry
     const defaults = {
         image: "freedom.png",
         title: "Enter Title",
@@ -88,6 +17,7 @@ function EntryCard(props){
         description: "Le description.",
     };
 
+    // Manages every entries values
     const [entries, setEntries] = useState(
         new Array(props.size).fill(null).map(() => ({
             image: defaults.image,
@@ -116,12 +46,13 @@ function EntryCard(props){
     const date = entries[keyValue]["date"];
     const description = entries[keyValue]["description"];
 
-    // Verify if the user is allowed to edit timeline entries.
+    // Verify if the user is allowed to edit timeline entries
     const [userCanEdit, setUserCanEdit] = useState(currentView === "edit-mode");
 
-    // Clear data
+    // Clear entry data
     useEffect(() => {
         if (deleteState){
+
             const newEntries = [...entries];
             newEntries[keyValue] = {
                 image: defaults.image,
@@ -129,12 +60,13 @@ function EntryCard(props){
                 date: defaults.date,
                 description: defaults.description,
             }
+
             setEntries(newEntries);
             setDeleteState(false);
         }
     }, [deleteState]);
 
-    // Disable editing abilities once currentView changes.
+    // Determine if user can edit timeline entry
     useEffect(() => {
         setUserCanEdit(currentView === "edit-mode");
     }, [currentView]);
@@ -157,39 +89,11 @@ function EntryCard(props){
     );
 }
 
+
 // PropTypes setup
 EntryCard.propTypes = {
     size: PropTypes.number.isRequired,
 };
-
-EntryImage.propTypes = {
-    image: PropTypes.string,
-    title: PropTypes.string,
-    index: PropTypes.number,
-    update: PropTypes.func,
-    disabled: PropTypes.bool,
-}
-
-EntryTitle.propTypes = {
-    title: PropTypes.string,
-    index: PropTypes.number,
-    update: PropTypes.func,
-    disabled: PropTypes.bool,
-}
-
-EntryDate.propTypes = {
-    date: PropTypes.string,
-    index: PropTypes.number,
-    update: PropTypes.func,
-    disabled: PropTypes.bool,
-}
-
-EntryDescription.propTypes = {
-    description: PropTypes.string,
-    index: PropTypes.number,
-    update: PropTypes.func,
-    disabled: PropTypes.bool,
-}
 
 // Exports
 export default EntryCard
