+++ src/index.css
@@ -1,3 +1,10 @@
+@import './Application/Application.css';
+@import './Header/Header.css';
+@import './TimelineGrid/TimelineGrid.css';
+@import './EntryCard/EntryCard.css';
+@import './EntryCard/EntryCardDescription.css';
+@import './DeleteEntryModal/DeleteEntryModal.css';
+
 :root {
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
